+++ generators/server/index.js
@@ -10,7 +10,17 @@ module.exports = class Server extends Generator {
   }
 
   install () {
-    this.npmInstall(['nodemon@latest'], { saveDev: true })
+    this.npmInstall(
+      [
+        'nodemon@latest',
+        '@types/body-parser@latest',
+        '@types/debug@latest',
+        '@types/express@latest',
+        '@types/morgan@latest',
+        '@types/node@latest'
+      ],
+      { saveDev: true }
+    )
 
     this.npmInstall([
       'body-parser@latest',